review routes finalized and tested

+++ app/api/reviews_routes.py
@@ -24,7 +24,9 @@ def review_by_id(id):
     """
 
     review = Review.query.get(id)
-    return review.to_dict()
+    if review:
+        return review.to_dict()
+    return 'OOps! Looks like this review does not exist', 404
 
 # get all reviews of the current user
 @review_routes.route('/current')
@@ -37,31 +39,6 @@ def current_user_reviews():
     reviews =  Review.query.filter_by(user_id = id)
 
     return [review.to_dict() for review in reviews]
-
-# # get all reviews of an album
-# @review_routes.route('albums/<int:id>')
-# @login_required
-# def album_reviews(id):
-#     """
-#     Query for all reviews for an album and returns them in a list of dictionaries
-#     """
-
-#     reviews =  Review.query.filter_by(album_id = id)
-
-#     return [review.to_dict() for review in reviews]
-
-
-# # get all reviews for a song
-# @review_routes.route('songs/<int:id>')
-# @login_required
-# def song_reviews(id):
-#     """
-#     Query for all reviews for a song and returns them in a list of dictionaries
-#     """
-
-#     reviews =  Review.query.filter_by(song_id = id)
-
-#     return [review.to_dict() for review in reviews]
 
 
 # # create a review for an album
@@ -77,7 +54,6 @@ def create_album_review(id):
         review = Review(
             user_id = session['_user_id'],
             reviewable_type = "Album",
-            reviewable_id = id,
             rating = form.data["rating"],
             comment = form.data["comment"],
             album_id = id
@@ -86,29 +62,6 @@ def create_album_review(id):
         db.session.commit()
         return review.to_dict()
     return form.errors, 400
-
-# # # create a review for a song
-# @review_routes.route('/songs/<int:id>', methods=["POST"])
-# @login_required
-# def create_song_review(id):
-#     """
-#         Create a review for a song
-#     """
-#     form = ReviewForm()
-#     form["csrf_token"].data = request.cookies["csrf_token"]
-#     if form.validate_on_submit():
-#         review = Review(
-#             user_id = session['_user_id'],
-#             reviewable_type = "Song",
-#             reviewable_id = id,
-#             rating = form.data["rating"],
-#             comment = form.data["comment"],
-#             song_id = id
-#         )
-#         db.session.add(review)
-#         db.session.commit()
-#         return review.to_dict()
-#     return form.errors, 400
 
 # update a review
 @review_routes.route('/<int:id>', methods=["PUT"])
@@ -145,6 +98,9 @@ def delete_song_review(id):
         Create a review for a song
     """
     review = Review.query.get(id)
+    if review == None:
+        return 'OOps! Looks like this review does not exist', 404
+
     if review.user_id == int(session['_user_id']):
         db.session.delete(review)
         db.session.commit()
